Add files via upload
This is the part of the code that i think i have the most troubles with today. I would love to develop my API to create a video that has descriptions superimposed on the pictures within the video. Therefore. I generate a second copy of all the pictures with the descriptions on them and rename all the pictures again. Moreover, I deleted the old pictures without the descriptions. In the end, I just Ffmpeg to generate the video with the pictures that included the descriptions in it.