MongoDB.phase1
In MongoDB phase1, we are guided to learn about how to access database and search data to gain desired result.  We are guided to use MongoDB as our database and I create two functions (Generate_database() as well as Search()) in my code to populate the database as well as searching. My code is compatible in Python 3.x version and I will provide later update in how to update the database.